+++ src/scanner_util.h
@@ -31,4 +31,28 @@ char* refreshBuffer(char * bufferToRefresh, int newSize);
 void skipWhitespace(char * buffer, int * index);
 
 
+/*!
+    Extracts a word from a buffer and returns it as a new malloced string. The word starts at index and ends when before the first character in boundrychars. Index will be moved to the first character in boundrychars.
+    \param buffer A buffer to read.
+    \param index A pointer to an int that holds an index in the buffer from which to start and tha twill be moved to the value of the end of the word extracted
+    \param boundrychars A set of characters that are boundries and end a word
+    \param l_boundrychars The length of boundrychars.
+    \returns A new malloced string containing the extracted word.
+    \author klm127
+    \date 2/11/2023
+*/
+char* extractWord(char* buffer, int * index, char* boundrychars, int l_boundrychars);
+
+/*!
+    Determines if a char matches a char within charset.
+    \param test A character to test.
+    \param charset A charset to check.
+    \param l_charset The length of the charset.
+    \return short, 0 for not in charset, 1 if in charset.
+    \author klm127
+    \date 2/11/2023
+
+*/
+short charIn(char test, char* charset, int l_charset);
+
 #endif