+++ tests/scanner_util_test.c
@@ -39,6 +39,51 @@ void test_skipWhitespace(CuTest *tc) {
     CuAssertIntEquals_Msg(tc, "should be at 'b'", 1, buf[index] == 'b');
 }
 
+void test_charin(CuTest *tc) {
+
+    short actual = charIn('a', (char[2]){'a','b'}, 2);
+    CuAssertIntEquals(tc, 1, actual);
+    actual = charIn('a', (char[1]){'b'}, 1);
+    CuAssertIntEquals(tc, 0, actual);
+    char boundries[8] = {' ', ';', '+', '-', '/', '=', '*', '\t' };
+    int l = 8;
+    actual = charIn('\t', boundries, l);
+    CuAssertIntEquals(tc, 1, actual);
+    actual = charIn(' ', boundries, l);
+    CuAssertIntEquals(tc, 1, actual);
+    actual = charIn(';', boundries, l);
+    CuAssertIntEquals(tc, 1, actual);
+
+}
+
+void test_extractWord(CuTest *tc) {
+    char * boundrychars = "\t+-;* \0";
+    int l_boundry = 6;
+    char* buffer = "hello   buffer+three\tkeyword;a-b*c    ";
+
+    int index = 0;
+    char * hello = extractWord(buffer, &index, boundrychars, l_boundry);
+    CuAssertStrEquals(tc, "hello", hello);
+    free(hello);
+
+    index += 3;
+    char * wbuffer = extractWord(buffer, &index, boundrychars, l_boundry);
+    CuAssertStrEquals(tc, "buffer", wbuffer);
+    free(wbuffer);
+
+    index += 1;
+    char * wthree = extractWord(buffer, &index, boundrychars, l_boundry);
+    CuAssertStrEquals(tc, "three", wthree);
+    free(wthree);
+
+    index += 1;
+    char * wkeyword = extractWord(buffer, &index, boundrychars, l_boundry);
+    CuAssertStrEquals(tc, "keyword", wkeyword);
+    free(wkeyword);
+
+
+    
+}
 
 
 CuSuite *scannerUtilGetSuite()
@@ -46,6 +91,8 @@ CuSuite *scannerUtilGetSuite()
     CuSuite *suite = CuSuiteNew();
     SUITE_ADD_TEST(suite, test_doubleBuffer);
     SUITE_ADD_TEST(suite, test_refreshBuffer);
+    SUITE_ADD_TEST(suite, test_charin);
+    SUITE_ADD_TEST(suite, test_extractWord);
 
     return suite; 
 }